% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNeighborsDistance.R
\name{analyzeNeighborsDistance}
\alias{analyzeNeighborsDistance}
\title{Analyze the distances to the upstream and downstream neighbors, for different categories of orientations (SameStrand / OppositeStrand)}
\usage{
analyzeNeighborsDistance(
  GeneList,
  GeneNeighborhood = NULL,
  DistriTest = TRUE,
  GeneUniverse = NULL,
  confLevel = 0.95,
  nboot = 10000,
  CItype = "bca",
  ncores = NULL
)
}
\arguments{
\item{GeneList}{A character vector of "focus" genes to analyze}

\item{GeneNeighborhood}{A \code{\link{tibble}} obtained with the \code{\link{getGeneNeighborhood}} function or any data frame
 with the following columns:
\itemize{
  \item \code{GeneName}: Name of the focus gene
  \item \code{Upstream}: Name of the gene located upstream of the focus gene
  \item \code{Downstream}: Name of the gene located downstream of the focus gene
  \item \code{UpstreamClass}. Class of the Upstream gene (e.g. SameStrand / OppositeStrand). The class name must contain the string 'overlap' in case of overlap
  \item \code{DownstreamClass}. Class of the Downstream gene
  \item \code{UpstreamDistance}. Distance between the focus gene and the Upstream gene
  \item \code{DownstreamDistance}. Distance between the focus gene and the Downstream gene
}}

\item{DistriTest}{A logical (default is TRUE) indicating if the distribution of distances should be compared to the reference universe}

\item{GeneUniverse}{An optional character vector of genes in the universe. By default all genes in \code{GeneNeighborhood} are considered in the Universe.}

\item{confLevel}{Confidence level for the intervals on the mean and median}

\item{nboot}{Number of bootstrap replicates used to estimate confidence intervals of the mean and median}

\item{CItype}{Type of bootstrap confidence interval ("perc" for classical percentile or "bca" for bias-corrected and accelerated intervals). Passed to \code{\link[boot]{boot.ci}} function}

\item{ncores}{Number of processes for parallel computation (passed to \code{\link[boot]{boot}} function).}
}
\value{
A list with:
\itemize{
  \item distances. Upstream and downstream intergenic distances for non overlapping genes
  \item stats. Statistics on the upstream/downstream intergenic distances for non overlapping genes
}
}
\description{
Essentially a wrapper that does some prefiltering on gene list,
  applies \code{\link{statDistanceSide}} to both Upstream and Downstream distances and formats the results
}
\section{DETAILS}{

The function only returns the distances for the GeneList, not for the GeneUniverse.
To get the distances for the GeneUniverse, use the GeneUniverse as the GeneList and DistriTest=FALSE (see examples)

Overlapping genes are removed.
The function should thus return results for the following orientations:
\itemize{
  \item \code{SameStrand}: The upstream/downstream gene is on the same strand as the focus gene
  \item \code{OppositeStrand}: The upstream/downstream gene is on the opposite strand of the focus gene
}
}

\examples{
## Obtain gene neighborhood information:
  GeneNeighbors <- getGeneNeighborhood(Genegr)
## Get a (random) set of (100) genes:
  set.seed(123)
  randGenes <- sample(names(Genegr), 100)
## Extract all distances and compare them to reference genes:
  NDS <- analyzeNeighborsDistance(GeneList = randGenes,
                                  GeneNeighborhood = GeneNeighbors,
                                  nboot = 1e3,
                                  ncores = 2)
  NDS$stats
## Extract the distances for all non-overlapping genes:
  alldist <- analyzeNeighborsDistance(GeneList = names(Genegr),
                                      GeneNeighborhood = GeneNeighbors,
                                      DistriTest = FALSE,
                                      nboot = 1e3,
                                      ncores = 2)
## Some stats on these distances:
  alldist$stats

}
\seealso{
\code{\link{statDistanceSide}}
}
\author{
Pascal GP Martin
}
