% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeNeighborsOrientation.R
\name{AnalyzeNeighborsOrientation}
\alias{AnalyzeNeighborsOrientation}
\title{Analyze the orientation of gene neighbors for a set of pre-defined genes}
\usage{
AnalyzeNeighborsOrientation(GeneList, GeneNeighborhood = GeneNeighbors,
  colNames = c(GeneName = "GeneName", Upstream = "Upstream", Downstream =
  "Downstream", UpstreamClass = "UpstreamClass", DownstreamClass =
  "DownstreamClass"), keepOther = TRUE, EnrichTest = TRUE,
  GeneUniverse = NULL)
}
\arguments{
\item{GeneList}{A character vector of focus genes to analyze}

\item{GeneNeighborhood}{A tibble obtained with the GetGeneNeighborhood function. Can also be any data.frame with the relevant information.}

\item{colNames}{A named character vector of length 5 containing the names of the columns from \code{GeneNeighborhood} containing the following information (in this order):
\itemize{
  \item \code{GeneName}. Name of the focus gene
  \item \code{Upstream}. Name of the gene located upstream of the focus gene
  \item \code{Downstream}. Name of the gene located downstream of the focus gene
  \item \code{UpstreamClass}. Class of the Upstream gene
  \item \code{DownstreamClass}. Class of the Downstream gene
}}

\item{keepOther}{A logical (default to \code{TRUE}) indicating if the category "other" should be analyzed or not}

\item{EnrichTest}{A logical (default to \code{TRUE}) indicating if Enrichment test should be performed}

\item{GeneUniverse}{An optional character vector of genes in the universe. By default all genes in GeneNeighborhood are considered in the Universe.}
}
\value{
A \code{tibble} with orientation and distance for upstream/downstream gene.
 If \code{EnrichTest} is \code{TRUE}, the \code{tibble} also contains data for the universe and a p-value from a Fisher exact test evaluating the enrichment for each orientation in \code{GeneList}.
}
\description{
Analyze the orientation of gene neighbors (upstream/downstream genes) for a set of pre-defined focus genes.
 Gives the number and percentage of genes in each orientation.
 Evaluate the enrichment of the different orientations when using a reference gene universe.
}
\section{DETAILS}{

The function analyzes the following orientations:
\itemize{
  \item \code{SameStrand}. The upstream/downstream gene is on the same strand as the focus gene
  \item \code{OppositeStrand}. The upstream/downstream gene is on the opposite strand of the focus gene
  \item \code{OppositeOverlap}. The upstream/downstream gene is on the opposite strand of the focus gene and overlaps with the focus gene
  \item \code{SameOverlap}. The upstream/downstream gene is on the same strand of the focus gene and overlaps with the focus gene
  \item \code{other}. (only if \code{keepOther} is \code{TRUE}) Any other, more complex situation (e.g. multiple overlaps, focus gene contained within another gene, etc.)
}
}

\examples{
## Obtain gene neighborhood information:
  GeneNeighbors <- GetGeneNeighborhood(Genegr)
## get a random set of 100 genes:
  set.seed(123)
  randGenes <- names(Genegr)[sample.int(676, 100)]
## Analyze the orientation of their neighbors
  AnalyzeNeighborsOrientation(randGenes,
                              GeneNeighborhood = GeneNeighbors)
## Select a less random set of genes:
  isOpposite <- grepl("Opposite",GeneNeighbors$UpstreamClass)
  probs <- ifelse(isOpposite, 0.6/sum(isOpposite), 0.4/sum(!isOpposite))
  set.seed(123)
  lessrandGenes <- GeneNeighbors$GeneName[sample.int(676, 100, prob=probs)]
  AnalyzeNeighborsOrientation(lessrandGenes,
                              GeneNeighborhood = GeneNeighbors)

}
\seealso{
\code{\link{fisher.test}}
}
\author{
Pascal GP Martin
}
