% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCoverageAroundFeatures.R
\name{AnnotationCoverageAroundFeatures}
\alias{AnnotationCoverageAroundFeatures}
\title{Extract stranded annotation profiles at and around a set of features/intervals.}
\usage{
AnnotationCoverageAroundFeatures(annot, features = NULL,
  sidedist = 2000L, usePercent = FALSE, BinFeatures = TRUE, ...)
}
\arguments{
\item{annot}{A \code{\link{GRanges}} of (stranded) genomic annotations}

\item{features}{An optional \code{GRanges} or a character vector with a subset of annot names to extract profiles at. Defaults to NULL for all annot features.}

\item{sidedist}{Integer. Distance on each side of features at which to extract the profiles}

\item{usePercent}{Logical. Should annotation coverage be expressed as presence/absence (0/1)? (instead of counting the number of features aligning at each position)}

\item{BinFeatures}{Logical. Should \code{features} be binned? Default to TRUE and will be forced to TRUE if \code{features} have different width.}

\item{...}{additional arguments passed to \code{\link{BinFeatureProfiles}}}
}
\value{
A list with the following elements:
\itemize{
  \item{"Feature_Sense"}{Coverage of annot (possibly binned) at features, on the sense strand (i.e. the strand of the feature)}
  \item{"Feature_Antisense"}{Coverage of annot (possibly binned) at features, on the antisense strand (i.e. the opposite strand of features)}
  \item{"UpstreamBorder_Sense"}{Coverage of annot around (+/- sidedist) the upstream border of features (Typically the TSS), on the sense strand}
  \item{"UpstreamBorder_Antisense"}{Coverage of annot around (+/- sidedist) the upstream border of features, on the antisense strand}
  \item{"DownstreamBorder_Sense"}{Coverage of annot around (+/- sidedist) the downstream border of features (Typically the TES), on the sense strand}
  \item{"DownstreamBorder_Antisense"}{Coverage of annot around (+/- sidedist) the downstream border of features, on the antisense strand}
}
}
\description{
Extract stranded annotation profiles at and around a set of features/intervals.
}
\section{DETAILS}{

The \code{usePercent} option is intended to give a percentage of bases covered by annotations when performing average profiles from several features. \cr
When calculating a coverage with usePercent=TRUE any value > 1 is replaced by 1. \cr
This is equivalent to calculating \code{coverage(reduce(annot))} instead of \code{coverage(annot)}.
}

\examples{
## Extract the profiles around (+/-50bp) all genes. We bin the genes in 3 bins only.
AllProf <- AnnotationCoverageAroundFeatures(Genegr, sidedist = 50, usePercent = TRUE, nbins=3)

}
\seealso{
BinFeatureProfiles
}
\author{
Pascal GP Martin
}
