% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleProfiles.R
\name{assembleProfiles}
\alias{assembleProfiles}
\title{Assemble profiles from the results of the annotationCoverageAroundFeatures function}
\usage{
assembleProfiles(annotcovr, sideDist = NULL)
}
\arguments{
\item{annotcovr}{a list with 6 \code{\link[IRanges]{RleList}}s named:
\itemize{
  \item \code{Feature_Sense}: Coverage (possibly binned) on the body of features, on the sense strand (i.e. the strand of the feature)
  \item \code{Feature_Antisense}: Coverage (possibly binned) on the body of features, on the antisense strand (i.e. the opposite strand of features)
  \item \code{UpstreamBorder_Sense}: Coverage centered on the upstream border of features (Typically the TSS), on the sense strand
  \item \code{UpstreamBorder_Antisense}: Coverage centered on the the upstream border of features, on the antisense strand
  \item \code{DownstreamBorder_Sense}: Coverage centered on the downstream border of features (Typically the TES), on the sense strand
  \item \code{DownstreamBorder_Antisense}: Coverage centered on the the downstream border of features, on the antisense strand
}
as produced by the \code{\link{annotationCoverageAroundFeatures}} function}

\item{sideDist}{Integer (optional). Distance to keep on each side of the features. Defaults to the maximum distance availabe in \code{annotcovr}}
}
\value{
a list of 2 RleLists corresponding to:
 \itemize{
  \item \code{Profiles_Sense}: RleList of profiles for the sense strand
  \item \code{Profiles_Antisense}: RleList of profiles for the antisense strand
 }
}
\description{
Assemble profiles from the results of the annotationCoverageAroundFeatures function.   \cr
             For each gene/feature, the function creates a profile containing:
\enumerate{
  \item A region upstream of the feature ending with the upstream border of the feature (typically the TSS)
  \item A region covering the body of the feature (possibly binned to normalize for different features' sizes)
  \item A region downstream of the feature starting with the downstream border of the feature (typically the TES)
  }
}
\examples{
## Extract the profiles around (+/-50bp) the first 10 genes. We bin the genes in 3 bins only.
  top10Prof <- annotationCoverageAroundFeatures(Genegr, features=1:10,
                                                sidedist = 50,
                                                usePercent = TRUE,
                                                nbins=3)
## Assemble the profiles:
  Prof <- assembleProfiles(top10Prof)
}
\seealso{
\code{\link{annotationCoverageAroundFeatures}}
}
