% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinFeatureProfiles.R
\name{BinFeatureProfiles}
\alias{BinFeatureProfiles}
\title{Bin coverage profiles given as a matrix or RleList of profiles.}
\usage{
BinFeatureProfiles(FeatureProfiles, nbins = 100L, binwidth = NULL,
  aggregFUN = "mean", asMatrix = FALSE, ...)
}
\arguments{
\item{FeatureProfiles}{A \code{\link{matrix}} (with features in row and genomic positions in column) or \code{\link{RleList}}, typically obtained with the \code{\link[GeneNeighborhood]{profcomp}} function.}

\item{nbins}{Integer. Number of bins to define along each feature. Default value is 100 bins.}

\item{binwidth}{Integer. Size of the bins to define on the features. Default value is NULL and 100 bins are used instead.}

\item{aggregFUN}{A character string (or a function name), such as "median", "mean", etc., with the name of a function used to aggregate the data within the bins. Defaults to "mean".
A user-defined aggregFUN should take an \code{\link{Rle}} as an argument and return a single value.}

\item{asMatrix}{Logical (default to FALSE). If TRUE, the binned profiles are returned as a matrix (only possible if featrures have the same number of bins).}

\item{...}{Further arguments passed to aggregFUN.}
}
\value{
An \code{RleList} or a \code{matrix} (if \code{asMatrix} is TRUE) of binned profiles.
}
\description{
Bin coverage profiles given as a matrix or RleList of profiles.
}
\section{DETAILS}{

 The function removes features of length < nbins (or binwidth) and returns an message with the number of removed features.
 The remaining features are sliced into \code{nbins} (possibly different bin sizes for different features) or into bins of size \code{binwidth} (possibly different number of bins for different features).
 Then the data in each bins is aggregated using the aggregFUN function
}

\examples{
## Obtain coverage for all genes:
  covr <- GenomicRanges::coverage(Genegr)
## Select a (random) set of (100) genes
  set.seed(123)
  randGenes <- sample(names(Genegr), 100)
## Windows of interests covering the gene bodies + 50bp on each side
  woi <- Genegr[randGenes] +50
## Coverage on these windows of interest:
  profs <- profcomp(covr, woi)
## Make 10 bins:
  BinFeatureProfiles(profs, nbins = 10)
## Return a matrix instead:
  BinFeatureProfiles(profs[1:3], nbins=10, asMatrix = TRUE)
## Make bins of size 20bp:
  BinFeatureProfiles(profs, binwidth=20)
##use your own function to aggregate the data within the bins
  mymean <- function(x) {mean(as.numeric(x), trim=0.1)}
  BinFeatureProfiles(profs[1:3], aggregFUN=mymean, nbins=10, asMatrix = TRUE)

}
\author{
Pascal GP Martin
}
