% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTests.R
\name{coinIndep}
\alias{coinIndep}
\title{Test the independence of 2 sets of distances with the coin package
(only returns the p-value)}
\usage{
coinIndep(.x, .y)
}
\arguments{
\item{.x}{A data frame with distances for the gene set of interest.}

\item{.y}{A data frame with distances for the control/reference gene set.}
}
\value{
The p-value of the independence test from the coin package.
}
\description{
Test the independence of 2 sets of distances with the coin package
(only returns the p-value)
}
\section{DETAILS}{

Both \code{.x} and \code{.y} should contain at least the following columns:
\itemize{
    \item Distance. Distance in bp.
    \item GeneName. Name of the focus gene.
    \item GeneSet. Name of the Gene Set
}
}

\examples{
## Create some random distance data
  set.seed(123)
  mydistData <- data.frame(GeneName = stringi::stri_rand_strings(400, 5),
                           Distance = sample(1:5000, 400, replace=TRUE),
                           GeneSet = sample(c("TestSet", "RefSet"),
                                            400, replace= TRUE))
## Compute the p-value of the independence test
  coinIndep(.x = mydistData[mydistData$GeneSet == "TestSet",],
            .y = mydistData[mydistData$GeneSet == "RefSet",])
## If .x is empty, the function returns NA
  coinIndep(.x = mydistData[mydistData$GeneSet == "SomeRandomName",],
            .y = mydistData[mydistData$GeneSet == "RefSet",])
## If .y is empty, the function returns an error
\dontrun{
  coinIndep(.x = mydistData[mydistData$GeneSet == "TestSet",],
            .y = mydistData[mydistData$GeneSet == "SomeRandomName",])
        }

}
\seealso{
\code{\link[coin]{independence_test}}
}
\author{
Pascal GP Martin
}
