% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2Neighbors.R
\name{dist2Neighbors}
\alias{dist2Neighbors}
\title{Extract distances to the upstream and downstream neighbors}
\usage{
dist2Neighbors(
  GeneNeighborhood = NULL,
  geneset = NULL,
  genesetName = "GeneSet"
)
}
\arguments{
\item{GeneNeighborhood}{A \code{tibble} obtained with the
\code{\link{getGeneNeighborhood}} function.
Can also be any data.frame with the relevant information.}

\item{geneset}{Either a character vector of identifiers defining the genes
of interest or a (named) list of such character vectors.}

\item{genesetName}{Character string. Name of the gene set (only used if
\code{geneset} is a character vector).
If \code{geneset} is a list, the names of its elements
are used as names of the gene sets.}
}
\value{
A tibble with the following columns:
\itemize{
  \item GeneName. ID of the focus gene.
  \item Neighbor. ID of the gene neighbor
  \item Orientation. Orientation of the neighbor (same or opposite strand).
  \item Side. Upstream or Downstream.
  \item Distance in bp
  \item GeneSet. Name of the gene set
}
}
\description{
Extract distances to the upstream and downstream neighbors.
             This is essentially a wrapper around \code{getDistSide} with
             the possibility to use several gene sets and with
             extra checks on the gene set(s)
}
\examples{
## Obtain gene neighborhood information:
  GeneNeighbors <- getGeneNeighborhood(Genegr)
## Get a (random) set of (100) genes:
  set.seed(123)
  randGenes <- sample(names(Genegr), 100)
## Extract distances to the upstream/downstream neighbors for this set:
  distForRandGenes <- dist2Neighbors(GeneNeighborhood = GeneNeighbors,
                                     geneset = randGenes,
                                     genesetName = "RandomGenes")
## Or for this set of genes and for all genes (the 'gene universe'):
 myGeneSets <- list("RandomGenes" = randGenes,
                    "AllGenes" = GeneNeighbors$GeneName)
 distForGeneSets <- dist2Neighbors(GeneNeighbors,
                                   myGeneSets)

}
\seealso{
\code{\link{getGeneNeighborhood}}
         \code{\link{getDistSide}}
}
\author{
Pascal GM Martin
}
