% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distStats.R
\name{distStats}
\alias{distStats}
\title{Descriptive statistics on intergenic distance data}
\usage{
distStats(GeneSetDistances, confLevel = 0.95, nboot = 10000,
  CItype = c("bca", "perc"), ncores = NULL)
}
\arguments{
\item{GeneSetDistances}{A \code{tibble} with intergenic distances for the
different gene sets, as generated by the
\code{\link{dist2Neighbors}} function.}

\item{confLevel}{Confidence level between 0 and 1.}

\item{nboot}{Integer (Default to 1e4). Number of bootstrap samples.}

\item{CItype}{Type of confidence interval ("bca" or "perc").}

\item{ncores}{Number of cores to use for boostrap.}
}
\value{
A tibble with the following columns:
\itemize{
  \item GeneSet. Name of the gene set.
  \item Side. Upstream or Downstream.
  \item Orientation. Orientation of the neighbor (SameStrand or OppositeStrand).
  \item n. Number of observations.
  \item Min. Minimum distance.
  \item Q1. First quartile.
  \item Median.
  \item Median_LowerCI. Lower bound of the median bootstrap confidence interval.
  \item Median_UpperCI. Upper bound of the median bootstrap confidence interval.
  \item Mean.
  \item Mean_LowerCI. Lower bound of the mean bootstrap confidence interval.
  \item Mean_UpperCI. Upper bound of the mean bootstrap confidence interval.
  \item Q3. Third quartile.
  \item Max. Maximum distance.
  \item SD. Standard deviation.
  \item SEM. Standard Error of the Mean.
}
}
\description{
Descriptive statistics on intergenic distance data
}
\section{NOTE}{

When nboot is too small and CItype=="bca", the following error can occur:
"estimated adjustment 'a' is NA"
In this case, use CItype="perc" and/or increase nboot.
Using nboot >= 1e4 (the default) is always recommended.
}

\examples{
## Obtain gene neighborhood information:
  GeneNeighbors <- getGeneNeighborhood(Genegr)
## Get a (random) set of (100) genes:
  set.seed(123)
  randGenes <- sample(names(Genegr), 100)
## Extract distances for this set of genes and for all genes :
 myGeneSets <- list("RandomGenes" = randGenes,
                    "AllGenes" = GeneNeighbors$GeneName)
 distForGeneSets <- dist2Neighbors(GeneNeighbors,
                                   myGeneSets)
## Obtain descriptive statistics for these sets of genes:
distStats(distForGeneSets,
          nboot = 1e3, CItype = "perc", ncores = 1)
}
\seealso{
\code{\link[boot]{boot}},
         \code{\link[boot]{boot.ci}}
}
