% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendPointPresence.R
\name{extendPointPresence}
\alias{extendPointPresence}
\title{Extend the presence value (1) of the closest single-point annotation to subsequent positions.}
\usage{
extendPointPresence(annotprofiles, sidedist = 2000L)
}
\arguments{
\item{annotprofiles}{An RleList similar to the one produced by the \code{\link{annotationCoverageAroundFeatures}} function.}

\item{sidedist}{an integer distance (in bp) around borders}
}
\value{
a list of 4 matrices with (extended) presence profiles at:
 \itemize{
 \item \code{upS}: the upstream border on the same strand ("sense")
 \item \code{upAS}: the upstream border on the opposite strand ("antisense")
 \item \code{dnS}: the downstream border on the same strand ("sense")
 \item \code{dnAS}: the downstream border on the opposite strand ("antisense")
 }
}
\description{
Single-point annotations are annotations that cover only 1 base in the genome.
             For example, a gene border such as the TSS or TES.
             To study the neighborhood of genes/features in terms of single-point annotation,
             we first extract annotation profiles using \code{\link{annotationCoverageAroundFeatures}} (with \code{usePercent=TRUE}).
             Then we use this function \code{extendPointPresence} which, for each gene/feature,
             defines the location of the closest single-point annotation at each border and for each strand
             and extends its presence to all subsequent positions.
}
\examples{
## Get the TSS of all genes:
  tss <- GenomicRanges::promoters(Genegr, upstream = 0, downstream = 1)
## Get their (presence/absence) coverage around (+/-50bp) genes (with 3 bins/gene):
  tsscov <- annotationCoverageAroundFeatures(annot = tss,
                                             features = Genegr,
                                             sidedist = 50,
                                             usePercent = TRUE,
                                             nbins = 3)
## For each gene, find the closest TSS and extend its presence to subsequent positions:
  extTSScov <- extendPointPresence(tsscov, sidedist=50)

}
\author{
Pascal GP Martin
}
