% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genegr.R
\docType{data}
\name{Genegr}
\alias{Genegr}
\title{A GRanges with 676 random genes on a single chromosome}
\format{A \code{\link{GRanges}} (from the \code{GenomicRanges} package).}
\usage{
Genegr
}
\description{
A \code{GRanges} with 676 genes named AA to ZZ with random start coordinate in 1:1e4,
random width in 1:10 and random strand on a single chromosome 'Chr1' of length 10010. Used for examples.
}
\author{
Pascal GP Martin
}
\keyword{data}
