% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNeighborhood.R
\name{getGeneNeighborhood}
\alias{getGeneNeighborhood}
\title{Extract info on upstream and downstream features/genes}
\usage{
getGeneNeighborhood(GeneGRanges)
}
\arguments{
\item{GeneGRanges}{A \code{GRanges} of feature/gene annotations}
}
\value{
A \code{\link{tibble}} with the following columns:
\itemize{
  \item \code{GeneName}: Name of the focus gene
  \item \code{Chr}: Seqnames of the focus gene
  \item \code{Strand}: Strand of the focus gene
  \item \code{Precede}: Name of the gene that the focus gene precedes (ignoring strand info)
  \item \code{Follow}: Name of the gene that the focus gene follows (ignoring strand info)
  \item \code{Upstream}: Name of the gene located upstream of the focus gene
  \item \code{Downstream}: Name of the gene located downstream of the focus gene
  \item \code{StrandUpstream}: Strand of the gene located upstream of the focus gene
  \item \code{StrandDownstream}: Strand of the gene located downstream of the focus gene
  \item \code{UpstreamOrientation}: Orientation of the upstream gene relative to the focus gene (S=same strand, O=Opposite strand or other)
  \item \code{DownstreamOrientation}: Orientation of the downstream gene relative to the focus gene (S, O or other)
  \item \code{UpstreamDistance}: Distance between the focus gene and its upstream neighbor
  \item \code{DownstreamDistance}: Distance between the focus gene and its downstream neighbor
  \item \code{CountOvlAnyStrand}: Number of genes with which the focus gene overlaps
  \item \code{ovlType}: Type of overlaps (see DETAILS)
  \item \code{ovlGene}: Name of the gene with which the focus gene overlaps (when it overlaps with a single gene)
  \item \code{NeighborClass}: Class of gene neighborhood (see DETAILS)
  \item \code{UpstreamClass}: Class of upstream Neighborhood (takes into account the presence of an overlap)
  \item \code{DownstreamClass}: Class of downstream Neighborhood (takes into account the presence of an overlap)
  \item \code{GenePair}: Type of gene pair formed between the focus gene and the gene that follows (see DETAILS)
  \item \code{Distance2Pair}: Intergenic distance between the focus gene and the gene that follows
}
}
\description{
Get orientation and distance info on the upstream and downstream features/genes from a GRanges
}
\section{DETAILS}{

Up/DownstreamOrientation columns do not precise if the up/downstream gene overlaps with the focus gene.

A distance of 0 in Up/DownstreamDistance columns can indicate that the genes overlap or that they are adjacent.

The type of overlaps in the ovlType column are coded as follow:
\itemize{
  \item c. The focus gene overlaps with more than 1 gene ("complex" overlap pattern)
  \item I. The focus gene shares the same borders (start AND end) with a gene on the same strand
  \item i. The focus gene shares the same borders (start AND end) with a gene on the opposite strand
  \item H. The focus gene "hosts" (=contains entirely) a gene on the same strand
  \item h. The focus gene "hosts" (=contains entirely) a gene on the opposite strand
  \item G. The focus gene is the "guest" of (=is entirely contained within) a gene that is on the same strand
  \item g. The focus gene is the "guest" of (=is entirely contained within) a gene that is on the opposite strand
  \item U. The 5' end of the focus gene (TSS) overlaps with a gene that is on the same strand
  \item u. The 5' end of the focus gene (TSS) overlaps with a gene that is on the opposite strand
  \item D. The 3' end of the focus gene (TES) overlaps with a gene that is on the same strand
  \item d. The 3' end of the focus gene (TES) overlaps with a gene that is on the opposite strand
}

 The Neighborclass column is defined as follow:
\itemize{
  \item{In the absence of overlap:}{ A 2-letter code using S (Same Strand) and O (Opposite Strand) to indicate the orientation of the uptstream (first letter) and downstream genes (second letter)}
  \item{When ovlType is u/U or d/D:}{ ovlType is combined with the information on the downstream (resp. upstream) gene to form the code (e.g. uS or uO)}
  \item{In all other cases:}{ NeighborClass is equal to ovlType}
}

 The codes in the GenePair column are defined as follow:
\itemize{
  \item{H2H:}{ Head-to-Head orientation: <---   --->}
  \item{T2H:}{ Tail-to-Head orientation: --->   --->}
  \item{T2T:}{ Tail-to-Tail orientation: --->   <---}
}
}

\examples{
\dontrun{
library(TxDb.Athaliana.BioMart.plantsmart25)
GeneNeighbors <- getGeneNeighborhood(GenomicFeatures::genes(TxDb.Athaliana.BioMart.plantsmart25))
# There are 155 genes (0.461\%) that overlap with >1 gene
table(GeneNeighbors$NeighborClass)
#
#   c    g    G    h    H    i    I   Od   OD   OO   OS   Sd   SD   SO   SS   uO   UO   uS   US
# 155  257  137  190   92    4   64 1448   52 6203 7461 1415   65 6237 9358  152   37  196   65
table(table(GeneNeighbors$GenePair))
#
#  H2H   T2H   T2T
# 7762 17173  7754
}


set.seed(123)
#The package includes a \\code{GRanges} named Genegr with 676 random genes on a single chromosome:
  Genegr

#Extract info on the neighbors of these genes:
  GeneNeighbors <- getGeneNeighborhood(Genegr)

#Classes of neighborhoods:
  table(GeneNeighbors$NeighborClass)
#Classes of gene pairs:
  table(GeneNeighbors$GenePair)

#Up/DownstreamOrientation columns don't document overlaps of the upstream/downstream gene:
  table(GeneNeighbors$NeighborClass[GeneNeighbors$UpstreamOrientation=="O"])
#This information is present in the Up/DownstreamClass columns:
  table(GeneNeighbors$NeighborClass[GeneNeighbors$DownstreamClass=="OppositeOverlap"])

}
\author{
Pascal GP Martin
}
