% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTES}
\alias{getTES}
\title{Extract the TES (=end of the gene) from a \code{\link{GRanges}}}
\usage{
getTES(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges}}.}
}
\value{
A GRanges of the TES positions
}
\description{
Extract the TES (=end of the gene) from a \code{\link{GRanges}}
}
\details{
The TES is defined as start(gr) when strand(gr)=="+" and
as end(gr) when strand(gr)=="-"
The function considers that strand=="*" is equivalent to strand=="+"
}
\examples{
## Get the TSS of genes in GeneGR:
  GenomicRanges::promoters(Genegr, upstream = 0, downstream = 1)
## Get their TES:
  getTES(Genegr)

}
\author{
Pascal GP Martin
}
