% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isBinaryMat}
\alias{isBinaryMat}
\title{Test if a matrix is a binary matrix}
\usage{
isBinaryMat(mat)
}
\arguments{
\item{mat}{A matrix. If mat is a data frame it will be converted by data.matrix with a warning.}
}
\value{
A logical. TRUE if the matrix is binary, FALSE otherwise
}
\description{
Test if a matrix is a binary matrix (i.e. contains only 0s and 1s (with possibly NA).
             Adapted from https://stackoverflow.com/a/23276062
}
\examples{
set.seed(123)
x <- matrix(rnorm(100), ncol= 10) # not a binary matrix
y <- matrix(sample(0:1, 100, rep=TRUE), ncol=10) # a binary matrix
z <- y ; z[x>1.5] <- NA #a binary matrix with some NA
isBinaryMat(x) #FALSE
isBinaryMat(y) #TRUE
isBinaryMat(z) #TRUE
}
