% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTests.R
\name{ksfun}
\alias{ksfun}
\title{Kolmogorov-Smirnov test on genomic distances}
\usage{
ksfun(.x, .y)
}
\arguments{
\item{.x}{A data frame with distances for the gene set of interest.}

\item{.y}{A data frame with distances for the control/reference gene set.}
}
\value{
The p-value of the Kolmorovov-Smirnov test.
}
\description{
Compare 2 sets of distances with a Kolmogorov-Smirnov test
(only returns the p-value)
}
\section{DETAILS}{

Both \code{.x} and \code{.y} should contain at least the following columns:
\itemize{
    \item Distance. Distance in bp.
    \item GeneName. Name of the focus gene.
}
Note that the Kolmogorov-Smirnov test is not adapted to integer values such
as intergenic distances in bp. However, it gives conservative p-values for
large enough gene sets.
}

\examples{
## Create some random distance data
  set.seed(123)
  mydistData <- data.frame(GeneName = stringi::stri_rand_strings(400, 5),
                           Distance = sample(1:5000, 400, replace=TRUE),
                           GeneSet = sample(c("TestSet", "RefSet"),
                                            400, replace= TRUE))
## Compute the p-value of the KS test comparing TestSet to RefSet
  ksfun(.x = mydistData[mydistData$GeneSet == "TestSet",],
        .y = mydistData[mydistData$GeneSet == "RefSet",])
## If .x is empty, the function returns NA
  ksfun(.x = mydistData[mydistData$GeneSet == "SomeRandomName",],
        .y = mydistData[mydistData$GeneSet == "RefSet",])
## If .y is empty, the function returns an error
\dontrun{
  ksfun(.x = mydistData[mydistData$GeneSet == "TestSet",],
        .y = mydistData[mydistData$GeneSet == "SomeRandomName",])
        }

}
\seealso{
\code{\link[stats]{ks.test}}
}
\author{
Pascal GP Martin
}
