% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCumulPercentProfile.R
\name{plotCumulPercentProfile}
\alias{plotCumulPercentProfile}
\title{Plot the cumulative proportions obtained from \code{\link{getCumulPercentProfiles}}}
\usage{
plotCumulPercentProfile(cumulprofs, gscolors = NULL, freeY = FALSE)
}
\arguments{
\item{cumulprofs}{A data frame with cumulative percentage profiles for different gene sets. As produced by the \code{\link{getCumulPercentProfiles}} function.}

\item{gscolors}{Optional named character vector with colors per gene sets}

\item{freeY}{Logical (defaults to FALSE) indicating if y-axis scales for panels/facets should be different}
}
\value{
A ggplot object
}
\description{
Plot the cumulative proportions obtained from \code{\link{getCumulPercentProfiles}}
}
\seealso{
\code{\link{getCumulPercentProfiles}}

#' @examples
## Get the TES of all genes:
  tes <- getTES(Genegr)
## Get their (presence/absence) coverage around (+/-50bp) genes (with 3 bins/gene):
  tescov <- annotationCoverageAroundFeatures(annot = tes,
                                             features = Genegr,
                                             sidedist = 50,
                                             usePercent = TRUE,
                                             nbins = 3)
## For each gene, find the closest TES and extend its presence to subsequent positions:
  extTEScov <- extendPointPresence(tescov, sidedist=50)
## Identify genes with a TES at less than 6bp from their TES, on any strand
  require(GenomicRanges)
  Dist2TES <- mcols(distanceToNearest(tes, ignore.strand = TRUE))$distance
  CloseTES <- names(Genegr)[Dist2TES<6]
  NotCloseTES <- names(Genegr)[Dist2TES>=6] #Also get the complementary set
## Calculate the cumulative percentage as the distance increases:
  CP <- getCumulPercentProfiles(extTEScov,
                                list("All" = names(Genegr),
                                     "CloseTES" = CloseTES,
                                     "FarTES" = NotCloseTES))
## Plot
  plotCumulPercentProfile(CP)
}
\author{
Pascal GP Martin
}
