% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNeighborsOrientation.R
\name{plotNeighborsOrientation}
\alias{plotNeighborsOrientation}
\title{Stacked barplot of gene orientations for upstream and downstream genes.}
\usage{
plotNeighborsOrientation(orientationStats)
}
\arguments{
\item{orientationStats}{A data frame produced by the \code{\link{analyzeNeighborsOrientation}} function}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plots stacked bars of the percentage of each orientation for the upstream and downstream genes.
}
\details{
The function only plots the following orientations, if present:
\itemize{
  \item SameStrand. The upstream/downstream gene is on the same strand as the focus gene
  \item OppositeStrand. The upstream/downstream gene is on the opposite strand of the focus gene
  \item OppositeOverlap. The upstream/downstream gene is on the opposite strand of the focus gene and overlaps with the focus gene
  \item SameOverlap. The upstream/downstream gene is on the same strand of the focus gene and overlaps with the focus gene
  \item other. Any other, more complex situation (e.g. multiple overlaps, focus gene contained within another gene, etc.)
}
}
\examples{
## Obtain gene neighborhood information:
  GeneNeighbors <- getGeneNeighborhood(Genegr)
## Define a (random) set of (100) genes:
  set.seed(123)
  randGenes <- names(Genegr)[sample.int(676, 100)]
## Analyze the orientation of their neighbors:
  NOS <- analyzeNeighborsOrientation(randGenes, GeneNeighborhood = GeneNeighbors)
## Plot the results:
  plotNeighborsOrientation(NOS)
}
\seealso{
\code{\link{ggplot2}}, \code{\link{ggplot}}
}
\author{
Pascal GP Martin
}
