% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profcomp}
\alias{profcomp}
\title{Extract windows of interest from a coverage.}
\usage{
profcomp(covr = NULL, woi = NULL)
}
\arguments{
\item{covr}{RleList. A genome-wide coverage (typically obtained with the \code{\link[GenomicRanges]{coverage}} function).}

\item{woi}{A \code{\link{GRanges}}. Windows of interest, over which to extract the profiles.}
}
\value{
An RleList with stranded profiles at woi
}
\description{
Extract profiles from a genome-wide coverage on a set of windows of interest. Reverses the profiles for windows that are on the '-' (minus) strand.
}
\section{DETAILS}{

Note that names of covr and seqnames of gr must match
}

\examples{
## Obtain coverage for all genes:
  covr <- GenomicRanges::coverage(Genegr)
## Select a (random) set of (100) genes
  set.seed(123)
  randGenes <- sample(names(Genegr), 100)
## Windows of interests cover the gene bodies + 50bp on each side
  woi <- Genegr[randGenes] + 50
## Coverage on these windows of interest:
  profcomp(covr, woi)


}
\author{
Pascal GP Martin
}
