% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTests.R
\name{pvalByGeneSet}
\alias{pvalByGeneSet}
\title{Apply a test to different genesets using the same reference/control set}
\usage{
pvalByGeneSet(tbdist, univ = "Universe", FUN, ...)
}
\arguments{
\item{tbdist}{A data frame with the following columns:
\itemize{
    \item Distance. Distance in bp.
    \item GeneName. Name of the focus gene.
    \item GeneSet. Name of the Gene Set
}}

\item{univ}{Character string giving the name of the control set
(should be in the GeneSet column).}

\item{FUN}{A function returning a single pvalue and that takes at least 2 arguments:
the data frame .x and .y, of the same form as \code{tbdist},
containing the distances for the test set and for the control/reference set.
and returns a single p-value.}

\item{...}{further arguments passed to FUN}
}
\value{
A tibble with 2 columns: GeneSet and pvalue
}
\description{
Apply a test to different genesets using the same reference/control set
}
\examples{
## Create some random distance data
  set.seed(123)
  mydistData <- data.frame(GeneName = stringi::stri_rand_strings(600, 5),
                           Distance = sample(1:5000, 600, replace=TRUE),
                           GeneSet = sample(c("TestSet1", "TestSet2", "RefSet"),
                                            600, replace= TRUE))
## Apply Kolmogorov-Smirnov test to the different test sets
  pvalByGeneSet(tbdist = mydistData,
                univ = "RefSet",
                FUN = ksfun)

}
\seealso{
\code{\link{replicate}}
\code{\link[matrixStats]{colMedians}}
}
\author{
Pascal GP Martin
}
