% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTests.R
\name{resampMed}
\alias{resampMed}
\title{Resampling test on the median (only returns the p-value)}
\usage{
resampMed(.x, .y, R = 10000)
}
\arguments{
\item{.x}{A data frame with distances for the gene set of interest.}

\item{.y}{A data frame with distances for the control/reference gene set.}

\item{R}{Number of samples to draw from \code{.y}}
}
\value{
A p-value representing the percentage of times that the median of
        \code{R} random samples drawn from the union of \code{.x} and \code{.y}
        is below the observed median of \code{.x}.
}
\description{
Resampling test on the median (only returns the p-value)
}
\section{DETAILS}{

Both \code{.x} and \code{.y} should contain at least the following columns:
\itemize{
    \item Distance. Distance in bp.
    \item GeneName. Name of the focus gene.
    \item GeneSet. Name of the Gene Set
}
}

\examples{
## Create some random distance data
  set.seed(123)
  mydistData <- data.frame(GeneName = stringi::stri_rand_strings(400, 5),
                           Distance = sample(1:5000, 400, replace=TRUE),
                           GeneSet = sample(c("TestSet", "RefSet"),
                                            400, replace= TRUE))
## Evaluate (by resampling) the probability to get a median lower that that of TestSet
## We use only 1000 random samples here for speed purposes but you should use >1e4
  resampMed(.x = mydistData[mydistData$GeneSet == "TestSet",],
            .y = mydistData[mydistData$GeneSet == "RefSet",],
            R = 1e3)
## If .x is empty, the function returns NA
  resampMed(.x = mydistData[mydistData$GeneSet == "SomeRandomName",],
            .y = mydistData[mydistData$GeneSet == "RefSet",],
            R = 1e3)
## If .y is empty, the function returns an error
\dontrun{
  resampMed(.x = mydistData[mydistData$GeneSet == "TestSet",],
            .y = mydistData[mydistData$GeneSet == "SomeRandomName",],
            R = 1e3)
        }

}
\seealso{
\code{\link{replicate}}
\code{\link[matrixStats]{colMedians}}
}
\author{
Pascal GP Martin
}
