% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RleList2matrix}
\alias{RleList2matrix}
\title{Convert an \code{RleList} to a \code{matrix}}
\usage{
RleList2matrix(rlelist = NULL)
}
\arguments{
\item{rlelist}{An \code{\link{RleList}}}
}
\value{
a \code{\link{matrix}} with one row for each element
        of the \code{RleList}
}
\description{
Convert an \code{RleList} to a \code{matrix}. The elements of
             the \code{RleList} should all have the same length.
}
\examples{
set.seed(123)
myrlelist <- IRanges::RleList(
                 x = S4Vectors::Rle(sample.int(100, 5)),
                 y = S4Vectors::Rle(sample.int(100, 5)),
                 z = S4Vectors::Rle(sample.int(100, 5)))
RleList2matrix(myrlelist)

}
\author{
Pascal GP Martin
}
