% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distTests.R
\name{Utest}
\alias{Utest}
\title{Compare 2 sets of distances with a Mann-Whitney U test
(only returns the p-value)}
\usage{
Utest(.x, .y)
}
\arguments{
\item{.x}{A data frame with distances for the gene set of interest.}

\item{.y}{A data frame with distances for the control/reference gene set.}
}
\value{
The p-value of the Mann-Whitney U test.
}
\description{
Compare 2 sets of distances with a Mann-Whitney U test
(only returns the p-value)
}
\section{DETAILS}{

Both \code{.x} and \code{.y} should contain at least the following columns:
\itemize{
    \item Distance. Distance in bp.
    \item GeneName. Name of the focus gene.
}
The Mann-Whithney U test is also called the Wilcoxon rank sum test.
}

\examples{
## Create some random distance data
  set.seed(123)
  mydistData <- data.frame(GeneName = stringi::stri_rand_strings(400, 5),
                           Distance = sample(1:5000, 400, replace=TRUE),
                           GeneSet = sample(c("TestSet", "RefSet"),
                                            400, replace= TRUE))
## Compute the p-value of the Mann-Whitney U-test comparing TestSet to RefSet
  Utest(.x = mydistData[mydistData$GeneSet == "TestSet",],
        .y = mydistData[mydistData$GeneSet == "RefSet",])
## If .x is empty, the function returns NA
  Utest(.x = mydistData[mydistData$GeneSet == "SomeRandomName",],
        .y = mydistData[mydistData$GeneSet == "RefSet",])
## If .y is empty, the function returns an error
\dontrun{
  Utest(.x = mydistData[mydistData$GeneSet == "TestSet",],
        .y = mydistData[mydistData$GeneSet == "SomeRandomName",])
        }

}
\seealso{
\code{\link[stats]{wilcox.test}}
}
