% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorizeMatrix.R
\name{winsorizeMatrix}
\alias{winsorizeMatrix}
\title{Winsorizing of an entire matrix}
\usage{
winsorizeMatrix(mat, low = 1, high = 99, verbose = FALSE)
}
\arguments{
\item{mat}{A matrix to winsorize}

\item{low}{Numeric. A value in [0,100]. All the values below \code{quantile(mat,low/100)} are replaced by \code{quantile(mat,low/100)}}

\item{high}{Numeric. A value in [0,100]. All the values above \code{quantile(mat,high/100)} are replaced by \code{quantile(mat,high/100)}}

\item{verbose}{Logical. Default to FALSE. If \code{verbose=TRUE}, the function will print statistics on the number of values replaced.}
}
\value{
A matrix with extreme values replaced by the corresponding quantiles.
}
\description{
Winsorizing of an entire matrix. Here, winsorization consists in replacing extreme values in a matrix by the corresponding upper and lower quantiles.
}
\examples{
## Create a matrix with random number from N(0,1):
  set.seed(123)
  x <- matrix(rnorm(1e4), ncol=100)
## Replace the top and bottom 1\% extreme values by the corresponding quantiles:
  y <- winsorizeMatrix(x, low=1, high=99, verbose = TRUE)
## Add some NA in x:
  x[x>2] <- NA
## winsorizeMatrix still works
  y <- winsorizeMatrix(x, low=1, high=99, verbose = TRUE)
}
