% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTES}
\alias{getTES}
\title{Extract the TES (=end of the gene/transcript) from a \code{\link{GRanges}} or \code{\link{GRangesList}}}
\usage{
getTES(gr)
}
\arguments{
\item{gr}{Either a \code{\link{GRanges}} or a \code{\link{GRangesList}} object.}
}
\value{
A GRanges or GRangesList of the TES positions
}
\description{
Extract the TES (=end of the gene/transcript) from a \code{\link{GRanges}} or \code{\link{GRangesList}}
}
\details{
The TES is defined as start(gr) when strand(gr)=="+" and
as end(gr) when strand(gr)=="-"
The function considers that strand=="*" is equivalent to strand=="+"
}
\examples{
## Get the TSS of genes in GeneGR:
  GenomicRanges::promoters(Genegr, upstream = 0, downstream = 1)
## Get their TES:
  getTES(Genegr)
## The function also works on GRangesList:
mygrglist <- S4Vectors::split(Genegr,
                              rep(paste0("gene", 1:169), each = 4))
getTES(mygrglist)

}
\author{
Pascal GP Martin
}
